Moved adgem to conftest
